refactor(core-json-rpc): reduce complexity and add types
