+++ __tests__/integration/core-json-rpc/blocks.test.ts
@@ -1,6 +1,7 @@
+import "jest-extended";
+
 import { app } from "@arkecosystem/core-container";
 import { Peer } from "@arkecosystem/core-p2p/src/peer";
-import "jest-extended";
 import nock from "nock";
 import { sendRequest } from "./__support__/request";
 import { setUp, tearDown } from "./__support__/setup";
