+++ __tests__/integration/core-json-rpc/transactions.test.ts
@@ -1,7 +1,8 @@
+import "jest-extended";
+
 import { app } from "@arkecosystem/core-container";
 import { Peer } from "@arkecosystem/core-p2p";
 import { Crypto } from "@arkecosystem/crypto";
-import "jest-extended";
 import nock from "nock";
 import { sendRequest } from "./__support__/request";
 import { setUp, tearDown } from "./__support__/setup";
