+++ packages/core-json-rpc/src/server/index.ts
@@ -1,7 +1,7 @@
 import { app } from "@arkecosystem/core-container";
 import { createServer, mountServer, plugins } from "@arkecosystem/core-http-utils";
 import { Logger } from "@arkecosystem/core-interfaces";
-import { registerMethods } from "./methods";
+import * as modules from "./modules";
 import { Processor } from "./services/processor";
 
 export async function startServer(options) {
@@ -29,13 +29,21 @@ export async function startServer(options) {
         });
     }
 
-    // @ts-ignore
-    registerMethods(server);
+    for (const module of Object.values(modules)) {
+        for (const method of Object.values(module)) {
+            // @ts-ignore
+            server.app.schemas[method.name] = method.schema;
+
+            delete method.schema;
+
+            server.method(method);
+        }
+    }
 
     server.route({
         method: "POST",
         path: "/",
-        async handler(request, h) {
+        async handler(request) {
             const processor = new Processor();
 
             return Array.isArray(request.payload)
