--- packages/core-json-rpc/src/server/methods/blocks/info.ts
@@ -1,18 +1,0 @@
-import Boom from "boom";
-import Joi from "joi";
-import { network } from "../../services/network";
-
-export const blockInfo = {
-    name: "blocks.info",
-    async method(params) {
-        const response = await network.sendRequest(`blocks/${params.id}`);
-
-        return response ? response.data : Boom.notFound(`Block ${params.id} could not be found.`);
-    },
-    schema: {
-        id: Joi.number()
-            // @ts-ignore
-            .unsafe()
-            .required(),
-    },
-};
