+++ packages/core-json-rpc/src/server/modules/index.ts
@@ -1,0 +1,5 @@
+import * as blocks from "./blocks";
+import * as transactions from "./transactions";
+import * as wallets from "./wallets";
+
+export { blocks, transactions, wallets };
