--- packages/core-json-rpc/src/server/utils/decrypt-wif.ts
@@ -1,14 +1,0 @@
-import { Crypto, Managers } from "@arkecosystem/crypto";
-import wif from "wif";
-
-export const decryptWIF = (encryptedWif, userId, bip38password) => {
-    const decrypted = Crypto.bip38.decrypt(encryptedWif.toString("hex"), bip38password + userId);
-
-    const encodedWIF = wif.encode(
-        Managers.configManager.get("network.wif"),
-        decrypted.privateKey,
-        decrypted.compressed,
-    );
-
-    return { keys: Crypto.crypto.getKeysFromWIF(encodedWIF), wif: encodedWIF };
-};
